requirements.txt
autogen-agentchat>=0.2.0
google-generativeai>=0.8.0
pandas>=2.0.0
pyyaml>=6.0
streamlit>=1.32.0
scikit-learn>=1.2.0
tiktoken>=0.4.0